pyproject.toml
[tool.poetry]
name = "cinema-api-auth"
version = "0.1.0"
description = ""
authors = ["AnasAbuHamideh <anas.abuhamida@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
